"Family" should be spelt with a lowercase 'f'
This caused a bug in production/preview whereby the capital F in family
deemed it to be a new section within rummager. This caused issues on the
browse pages as the link it would attempt to hit would be /browse/Family
instead of /browse/family.

Keep the slimmer headers lowercase.